Add files via upload
This is a class with a single method to create a player. It can be called from the class that takes in user input then added it to the players list in gameboard after the gameboard is created (if this is what we going to do). It has conflicts with the Player constructor due to parameters.